fix: moved db and auth into modules. + other small fixes
