+++ jest.config.ts
@@ -11,7 +11,7 @@ const defaultSettings = {
     '@ukef/config/(.*)': '<rootDir>/../src/config/$1',
     '@ukef/helpers/(.*)': '<rootDir>/../src/helpers/$1',
     '@ukef/module/(.*)': '<rootDir>/../src/modules/$1',
-    '@ukef/auth/(.*)': '<rootDir>/../src/auth/$1',
+    '@ukef/auth/(.*)': '<rootDir>/../src/modules/auth/$1',
     '@ukef/(.*)': '<rootDir>/../src/$1',
   },
 };
@@ -23,7 +23,7 @@ const config: JestConfigWithTsJest = {
       testMatch: ['**/*.test.ts'],
       transform: { '^.+\\.(ts|tsx)?$': ['ts-jest', { useESM: true }] },
       ...defaultSettings,
-      rootDir: 'src', // Since unit tests are created inside `src`
+      rootDir: 'src',
     },
     {
       displayName: 'API',
