+++ src/constants/auth.constant.ts
@@ -1,0 +1,3 @@
+export const AUTH = {
+  STRATEGY: 'x-api-key',
+};
