+++ src/constants/index.ts
@@ -4,7 +4,9 @@
  *
  * 1. Date
  * 2. Database name
+ * 3. Auth strategy
  */
 
+export * from './auth.constant';
 export * from './database-name.constant';
 export * from './date.constant';
