+++ src/main.module.ts
@@ -1,9 +1,6 @@
 import { Module } from '@nestjs/common';
 import { ConfigModule } from '@nestjs/config';
-import { AuthModule } from '@ukef/auth/auth.module';
-import { AuthService } from '@ukef/auth/auth.service';
 import config from '@ukef/config';
-import { DatabaseModule } from '@ukef/database/database.module';
 import { MdmModule } from '@ukef/module/mdm.module';
 import { LoggerModule } from 'nestjs-pino';
 
@@ -26,11 +23,7 @@ import { LoggerModule } from 'nestjs-pino';
         },
       },
     }),
-    AuthModule,
-    DatabaseModule,
     MdmModule,
   ],
-  controllers: [],
-  providers: [AuthService],
 })
 export class MainModule {}
