+++ src/modules/auth/auth.service.ts
@@ -3,9 +3,9 @@ import { ConfigService } from '@nestjs/config';
 
 @Injectable()
 export class AuthService {
-  constructor(private configService: ConfigService) {}
+  constructor(private readonly configService: ConfigService) {}
 
-  validateApiKey(key: string) {
+  validateApiKey(key: string): boolean {
     const apiKey: string = this.configService.get<string>('app.apiKey');
     return apiKey === key;
   }
