+++ src/modules/auth/guard/api-key.guard.ts
@@ -1,5 +1,6 @@
 import { Injectable } from '@nestjs/common';
 import { AuthGuard } from '@nestjs/passport';
+import { AUTH } from '@ukef/constants';
 
 @Injectable()
-export class ApiKeyAuthGuard extends AuthGuard('x-api-key') {}
+export class ApiKeyAuthGuard extends AuthGuard(AUTH.STRATEGY) {}
