+++ src/modules/auth/strategy/api-key.strategy.ts
@@ -1,18 +1,19 @@
 import { Injectable, UnauthorizedException } from '@nestjs/common';
 import { ConfigService } from '@nestjs/config';
 import { PassportStrategy } from '@nestjs/passport';
+import { AUTH } from '@ukef/constants';
 import { HeaderAPIKeyStrategy } from 'passport-headerapikey';
 
 import { AuthService } from '../auth.service';
 
 @Injectable()
-export class ApiKeyStrategy extends PassportStrategy(HeaderAPIKeyStrategy, 'x-api-key') {
-  constructor(private authService: AuthService, private configService: ConfigService) {
+export class ApiKeyStrategy extends PassportStrategy(HeaderAPIKeyStrategy, AUTH.STRATEGY) {
+  constructor(private readonly authService: AuthService, private readonly configService: ConfigService) {
     const headerKeyApiKey: string = configService.get<string>('app.apiKeyStrategy');
     super({ header: headerKeyApiKey, prefix: '' }, true, (apiKey: string, done: (arg0: UnauthorizedException, arg1: boolean) => void) => {
-      const hasValidKey: boolean = this.authService.validateApiKey(apiKey);
+      const hasValidKey = this.authService.validateApiKey(apiKey);
       if (hasValidKey) {
-        done(null, true);
+        return done(null, true);
       }
       done(new UnauthorizedException(), null);
     });
