+++ src/modules/mdm.module.ts
@@ -1,4 +1,6 @@
 import { Module } from '@nestjs/common';
+import { AuthModule } from '@ukef/auth/auth.module';
+import { DatabaseModule } from '@ukef/database/database.module';
 import { ConstantsModule } from '@ukef/module/constants/constants.module';
 import { CurrenciesModule } from '@ukef/module/currencies/currencies.module';
 import { ExposurePeriodModule } from '@ukef/module/exposure-period/exposure-period.module';
@@ -10,6 +12,8 @@ import { SectorIndustriesModule } from '@ukef/module/sector-industries/sector-in
 
 @Module({
   imports: [
+    AuthModule,
+    DatabaseModule,
     ConstantsModule,
     ExposurePeriodModule,
     HealthcheckModule,
@@ -20,6 +24,8 @@ import { SectorIndustriesModule } from '@ukef/module/sector-industries/sector-in
     CurrenciesModule,
   ],
   exports: [
+    AuthModule,
+    DatabaseModule,
     ConstantsModule,
     ExposurePeriodModule,
     HealthcheckModule,
