+++ tsconfig.json
@@ -30,10 +30,10 @@
       "@ukef/*": ["src/*"],
       "@ukef/constants/*": ["src/constants/index"],
       "@ukef/config/*": ["src/config/index"],
-      "@ukef/database/*": ["src/database/*"],
+      "@ukef/database/*": ["src/modules/database/*"],
       "@ukef/helpers/*": ["src/helpers/*"],
       "@ukef/module/*": ["src/modules/*"],
-      "@ukef/auth/*": ["src/auth/*"]
+      "@ukef/auth/*": ["src/modules/auth/*"]
     }
   },
   "include": ["src", "test", "jest.config.ts"],
